requirements.txt
django==5.0.1
djangorestframework==3.14.0
gunicorn==21.2.0